requirements.txt
ipykernel
streamlit
streamlit_drawable_canvas

numpy<2.0
opencv-python

matplotlib
tqdm                
pyyaml
imutils

torch
torchvision
doclayout_yolo
datasets

shapely 
pytesseract 
python-Levenshtein

fpdf2